# ranking: Optimize occurrence search in text